Move PasswordValidator to avoid circular import
    {apps.authentication.services => apps.users}.password_validation
